requirements.txt
Flask
mysql-connector-python
pytesseract
pdfplumber
Pillow
joblib
scikit-learn
python-dotenv
bcrypt
